+++ modules/canvaskit/skottie_bindings.cpp
@@ -13,8 +13,11 @@
 #include "modules/skottie/include/Skottie.h"
 #include "modules/skottie/include/SkottieProperty.h"
 #include "modules/skottie/utils/SkottieUtils.h"
+#include "modules/skottie/utils/TextEditor.h"
 #include "modules/skresources/include/SkResources.h"
 #include "modules/sksg/include/SkSGInvalidationController.h"
+#include "tools/skui/InputState.h"
+#include "tools/skui/ModifierKey.h"
 
 #include <string>
 #include <vector>
@@ -149,7 +152,7 @@ public:
                                         emscripten::val logger) {
         auto mgr = std::make_unique<skottie_utils::CustomPropertyManager>(
                         skottie_utils::CustomPropertyManager::Mode::kCollapseProperties,
-                        prop_prefix.empty() ? nullptr : prop_prefix.c_str());
+                        prop_prefix.c_str());
         static constexpr char kInterceptPrefix[] = "__";
         auto pinterceptor =
             sk_make_sp<skottie_utils::ExternalAnimationPrecompInterceptor>(rp, kInterceptPrefix);
@@ -352,6 +355,72 @@ public:
         return fSlotMgr->setVec2Slot(SkString(slotID), v);
     }
 
+    bool attachEditor(const std::string& layerID, size_t layerIndex) {
+        if (fTextEditor) {
+            fTextEditor->setEnabled(false);
+            fTextEditor = nullptr;
+        }
+
+        if (layerID.empty()) {
+            return true;
+        }
+
+        auto txt_handle = fPropMgr->getTextHandle(layerID, layerIndex);
+        if (!txt_handle) {
+            return false;
+        }
+
+        std::vector<std::unique_ptr<skottie::TextPropertyHandle>> deps;
+        for (size_t i = 0; ; ++i) {
+            if (i == layerIndex) {
+                continue;
+            }
+
+            auto dep_handle = fPropMgr->getTextHandle(layerID, i);
+            if (!dep_handle) {
+                break;
+            }
+            deps.push_back(std::move(dep_handle));
+        }
+
+        fTextEditor = sk_make_sp<skottie_utils::TextEditor>(std::move(txt_handle),
+                                                            std::move(deps));
+        return true;
+    }
+
+    void enableEditor(bool enable) {
+        if (fTextEditor) {
+            fTextEditor->setEnabled(enable);
+        }
+    }
+
+    bool dispatchEditorKey(const std::string& key) {
+        // Map some useful keys to the current (odd) text editor bindings.
+        // TODO: Add support for custom bindings in the editor.
+        auto key2char = [](const std::string& key) -> SkUnichar {
+            // Special keys.
+            if (key == "ArrowLeft")  return '[';
+            if (key == "ArrowRight") return ']';
+            if (key == "Backspace")  return '\\';
+
+            // Passthrough regular keys.
+            if (key.size() == 1) return key[0];
+
+            // Ignored.
+            return '\0';
+        };
+
+        return fTextEditor
+                ? fTextEditor->onCharInput(key2char(key))
+                : false;
+    }
+
+    bool dispatchEditorPointer(float x, float y, skui::InputState state, skui::ModifierKey mod) {
+        return fTextEditor
+                ? fTextEditor->onMouseInput(x, y, state, mod)
+                : false;
+    }
+
 private:
     ManagedAnimation(sk_sp<skottie::Animation> animation,
                      std::unique_ptr<skottie_utils::CustomPropertyManager> propMgr,
@@ -367,6 +436,8 @@ private:
     const std::unique_ptr<skottie_utils::CustomPropertyManager> fPropMgr;
     const sk_sp<skottie::SlotManager>                           fSlotMgr;
     const sk_sp<skresources::ResourceProvider>                  fResourceProvider;
+
+    sk_sp<skottie_utils::TextEditor>                            fTextEditor;
 };
 
 } // anonymous ns
@@ -462,7 +533,12 @@ EMSCRIPTEN_BINDINGS(Skottie) {
         }))
         .function("getScalarSlot"    , &ManagedAnimation::getScalarSlot)
         .function("setScalarSlot"    , &ManagedAnimation::setScalarSlot)
-        .function("setImageSlot"     , &ManagedAnimation::setImageSlot);
+        .function("setImageSlot"     , &ManagedAnimation::setImageSlot)
+
+        .function("attachEditor"         , &ManagedAnimation::attachEditor)
+        .function("enableEditor"         , &ManagedAnimation::enableEditor)
+        .function("dispatchEditorKey"    , &ManagedAnimation::dispatchEditorKey)
+        .function("dispatchEditorPointer", &ManagedAnimation::dispatchEditorPointer);
 
     function("_MakeManagedAnimation", optional_override([](std::string json,
                                                            size_t assetCount,
@@ -492,5 +568,21 @@ EMSCRIPTEN_BINDINGS(Skottie) {
                                                                      std::move(soundMap))),
                                       prop_prefix, std::move(logger));
     }));
+
+    enum_<skui::InputState>("InputState")
+        .value("Down",  skui::InputState::kDown)
+        .value("Up",    skui::InputState::kUp)
+        .value("Move",  skui::InputState::kMove)
+        .value("Right", skui::InputState::kRight)
+        .value("Left",  skui::InputState::kLeft);
+
+    enum_<skui::ModifierKey>("ModifierKey")
+        .value("None"      , skui::ModifierKey::kNone)
+        .value("Shift"     , skui::ModifierKey::kShift)
+        .value("Control"   , skui::ModifierKey::kControl)
+        .value("Option"    , skui::ModifierKey::kOption)
+        .value("Command"   , skui::ModifierKey::kCommand)
+        .value("FirstPress", skui::ModifierKey::kFirstPress);
+
     constant("managed_skottie", true);
 }